+++ lib/grape/exceptions/invalid_message_body.rb
@@ -3,7 +3,7 @@ module Grape
   module Exceptions
     class InvalidMessageBody < Base
       def initialize(body_format)
-        super(message: compose_message('invalid_message_body', body_format: body_format), status: 400)
+        super(message: compose_message(:invalid_message_body, body_format: body_format), status: 400)
       end
     end
   end