+++ lib/grape/exceptions/validation.rb
@@ -9,8 +9,7 @@ module Grape
       def initialize(args = {})
         fail 'Params are missing:' unless args.key? :params
         @params = args[:params]
-        @message_key = args[:message_key]
-        args[:message] = translate_message(args[:message_key]) if args.key? :message_key
+        args[:message] = translate_message(args[:message]) if args.key? :message
         super
       end