+++ lib/grape/validations/validators/values.rb
@@ -2,7 +2,7 @@ module Grape
   module Validations
     class ValuesValidator < Base
       def initialize(attrs, options, required, scope)
-        @values = options
+        @values = (options_key?(:value, options) ? options[:value] : options)
         super
       end
 
@@ -13,7 +13,7 @@ module Grape
         values = @values.is_a?(Proc) ? @values.call : @values
         param_array = params[attr_name].nil? ? [nil] : Array.wrap(params[attr_name])
         return if param_array.all? { |param| values.include?(param) }
-        fail Grape::Exceptions::Validation, params: [@scope.full_name(attr_name)], message_key: :values
+        fail Grape::Exceptions::Validation, params: [@scope.full_name(attr_name)], message: message(:values)
       end
 
       private